refactor: move FuzzConfig/FuzzWorkloadConfig to homeboy-extension-contract
Tenth sub-cut of the extension-contract campaign (Stage 1, wiki 12840).

Collects the payoff of the lifecycle-contract extraction (#8752): FuzzConfig
and FuzzWorkloadConfig were the last types stuck in manifest_config.rs, blocked
only because FuzzWorkloadConfig carries an Option<LifecycleContract> field.
LifecycleContract now lives in the leaf homeboy-lifecycle-contract crate, so
the extension-contract crate can depend on it and pull these two pure serde
types in cleanly.

manifest_config.rs is now fully drained of data types (only re-exports of
autofix_config/trace_config + a test remain). manifest.rs re-exports the two
Fuzz types from the contract crate, keeping crate::extension::* stable.
