+++ crates/homeboy-extension-contract/Cargo.toml
@@ -17,6 +17,7 @@ homeboy-error = { version = "0.1.0", path = "../homeboy-error" }
 homeboy-audit-contract = { version = "0.1.0", path = "../homeboy-audit-contract" }
 homeboy-product-identity = { version = "0.1.0", path = "../homeboy-product-identity" }
 homeboy-engine-primitives = { version = "0.1.0", path = "../homeboy-engine-primitives" }
+homeboy-lifecycle-contract = { version = "0.1.0", path = "../homeboy-lifecycle-contract" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 semver = "1.0"
